Convert project to ARC

+++ include/wintype.h
@@ -5,7 +5,7 @@
 #ifndef WINTYPE_H
 #define WINTYPE_H
 
-typedef long winid; /* a window identifier */
+typedef int winid; /* a window identifier */
 
 /* generic parameter - must not be any larger than a pointer */
 typedef union any {
